optimisation go brrr
